requirements.txt
aiofiles==22.1.0
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
Authlib==1.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
fastapi-jwt-auth==0.5.0
fastapi_amis_admin==0.3.0
graphql-core==3.2.3
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
prompt-toolkit==3.0.33
psycopg==3.1.2.dev0
psycopg-binary==3.1.4
psycopg-pool==3.1.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pyotp==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
redis==4.3.5
rfc3986==1.5.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlalchemy_database==0.0.11
sqlmodel==0.0.8
starlette==0.21.0
strawberry-graphql==0.142.2
typing_extensions==4.4.0
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
